Test for `composite: true` errors in CI, ensuring we can't ship those issues (#501)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

We've seen some issues crop up when `composite: true` is present in
`tsconfig.json` files. This PR adds a `composite: true` project where we
test the Inngest package for compliance for the types that it exports.

We purposefully want to limit the number of types exported from the main
`"inngest"` entrypoint, as each type exported here becomes part of the
public API, where changing those types is a breaking change. Therefore,
the composite check gives us a (albeit inaccurate) test that can stop us
shipping these dangerous changes.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix and testing
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- #384
- #385 
- #437
- #460