Add `typescript@rc` to test for `typescript@5.4` (#500)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Adds `typescript@rc` (currently `typescript@5.4.0`). This is used to
test the 5.4 release candidate to resolve any issues before it ships.
When it does, we'll move this PR to target `typescript@5.4.0` and merge,
including a changeset if any changes needed to be made.

We should ship this change once the final build is submitted at the
beginning of March. See microsoft/TypeScript#56948

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

- microsoft/TypeScript#56948