Allow `unknown` keys and add tests for custom `Jsonify` type
`Jsonify` is used to represent a step output going through serialization
as it passes to and from an Inngest Server in JSON. This used to be part
of the `type-fest` package, but we removed that when introducing support
for TypeScript 5.4 in #500, as our version support range for TS exceeded
what was provided by the `type-fest` package.

As part of that, we've pulled `Jsonify` into our code. There are some
issues (in the related section) we can now fix much more easily.

The issue that motivated this change (#509) would also have been solved
by allowing a developer to provide types that describe how the output of
a step is transformed. This is available, but is a rare ask and needs
time to implement.