INN-2861 No longer execute `step.sendEvent()` inline (#518)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

In INN-1349, `step.sendEvent()` became an "inline action." At the time,
these were intended to address situations where a developer misused
`step.sendEvent()`, assuming it was the only way to send an event, but
they didn't want to create a step function; the other actions in the
function were not wrapped in `step.run()`.

While this resolved some issues for early adopters, it created a
frustrating long-term issue is `step.sendEvent()` is the first action in
a run. An SDK has no way of knowing if it should be a step or not,
resulting in the potential that the event can be sent many times.

We should move back to the more explicit model of `step.*` methods
**always** creating a step function.

The recovery method outlined in the
[spec](https://github.com/inngest/inngest/blob/main/docs/SDK_SPEC.md)
(and present in v3+ of the TS SDK) mean that this will not break current
flows that rely on this behaviour.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [ ] ~Added unit/integration tests~ N/A Removing feature
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- INN-2861
- INN-1349