Fix event API errors not being parsed and returned (#535)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

When the event API responds with an error, we attempt to parse that
error and include it in the `Error` returned to the user.

The tests for this, however, mocked the incorrect response from Inngest,
resulting in the illusion of this functioning. The mock has been changed
and the fix implemented by slightly changing the parsing of the
response.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [x] Added unit/integration tests
- [x] Added changesets if applicable