INN-2882 Add selective header forwarding when sending events (#520)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Read the incoming headers tracestate and traceparent. If it is set, copy
the header to any outgoing event-sending requests.

This only covers `step.sendEvent()`; we'll add later support for
`client.send()` as we need async hooks for this which has scattered
runtime support.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A
- [ ] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- INN-2882