Fix `"inngest/hono"` serve handler not parsing environment variables (#571)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

The `"inngest/hono"` serve handler was not fetching environment
variables, which are not reliably accessible via globals in common
environments such as Cloudflare Workers.

In addition, `url` handling was a little wonky and didn't seem to handle
many variants. The hope was that Hono handled that to ensure that
`c.req.url` is always/never absolute, but this doesn't seem to be the
case.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A
- [ ] ~Added unit/integration tests~ N/A
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Supersedes #557 as we should no longer be required to manually set
signing and event keys
- Fixes #560's URL issues
- Fixes _some_ of the isues in #556, though there are still some
unknowns for the C++ errors there