Fix/incorrect custom fetch wrapping (#597)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

A warning log for when a custom fetch implementation is suspected to be
broken is a bit too spammy.

```
A request failed when using a custom fetch implementation; this may be a misconfiguration. Make sure that your fetch client is correctly bound to the global scope.
```

This log is intended to be shown when we suspect that the given fetch
implementation is invalid and seems to not be functioning correctly, for
example if it has lost its binding after being passed in an object.

At present, however, it is logged upon every failure, even if it is
correctly handled by code further up the stack.

This PR ensures we only log if it's not an expected fetch failure like
failing to reach a server, which still indicates that the implementation
is functional.

It also fixes an issue where the fetch implementation used by the
`InngestCommHandler` is wrapped and treated as custom if no custom
implementation has been given to `new Inngest()`.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A
- [x] Added unit/integration tests
- [x] Added changesets if applicable