Keep `^` specifier for `inngest` peer deps; changesets bug keeps us here
`changesets` can handle _not_ bumping the version of peer dependencies
when releasing so long as they still sit within the range. If we use
`>=`, it seems this experimental rule is bypassed and it still bumps,
causing our lock files to step out of sync.

For now, we'll go against the changes in #604 and keep this caret
specifier here.