Add support for global `Netlify.env` objects (#611)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Netlify, though it uses Deno under the hood, rename `Deno.env` to
`Netlify.env`. This PR adds support for this new global and ensures we
use it to access environment variables.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A
- [ ] ~Added unit/integration tests~ N/A
- [x] Added changesets if applicable