Fix rare checkpoint hang (#622)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Fixes a theoretical, rare issue whereby multiple checkpoints can be
reported before the next one is awaited, resulting in an entire
checkpoint being missed.

At best, this can cause the core loop to hang, and at worst make strange
decisions.

> [!NOTE]
> The current checkpoints that are supported mean that we never see this
issue. However, this PR guards us against altering the checkpoints in
the future and not accounting for this quirk.

For example:

- `await checkpointPromise` resolves
- `yield` reports checkpoint
- Checkpoint B is found; `setCheckpoint` is called, which reassigns
`checkpointPromise`
- Checkpoint C is found; `setCheckpoint` is called, which reassigns
`checkpointPromise`
- Core loop `while` ticks over and hits `await checkpointPromise`, which
resolves with Checkpoint C
- We've missed Checkpoint B

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A Internal fix
- [ ] ~Added unit/integration tests~ N/A Existing cover this fine; the
unit tests for this would be much too abstract
- [x] Added changesets if applicable