Loosen typing on `match` to remove performance concerns (#644)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Loosens typing of `match` in `step.waitForEvent()` and `cancelOn`
configurations in order to remove TypeScript performance concerns when
many complex events are used. It is now simply any `string` and is
unvalidated against event typing.

This supersedes #642 and marks `match` fields as deprecated; they will
be removed fully or reworked in v4 as we coalesce towards a solid
solution for writing matching expressions for these tools.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Supersedes #642
- Follows a previous attempt to retain the feature in #615