Make DefaultEncryptionService use libsodium (#555)
## Summary
Change `DefaultEncryptionService` in `@inngest/middleware-encryption` to
use libsodium. This is a breaking change.

Standardizing our SDKs' encryption middlewares on libsodium allows for
easy interoperability. For example, our TypeScript SDK could send an
encrypted event to our Python SDK.

## Todo

- [x] Make two middleware for proper support 🫠
- [x] Async encrypt/decrypt

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Documented in-package
- [x] Added unit/integration tests
- [x] Added changesets if applicable

---------

Co-authored-by: Jack Williams <1736957+jpwilliams@users.noreply.github.com>
Co-authored-by: Jack Williams <jack@inngest.com>