Add `inngest.setEnvVars(env)` to set env vars late on the client (#665)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

On some platforms, environment variables are sent as a function argument
and are not available as a global. `serve()` handlers manage this
transparently as they can silently retrieve the values they are passed,
but `Inngest` clients can require manual intervention.

For these cases, we add `inngest.setEnvVars()`, allowing you to set
environment variables at runtime and rest any URLs or configuration that
was set during instantiation.

```ts
app.on("GET", "/api/send-some-event", async (c) => {
  inngest.setEnvVars(c.env);

  await inngest.send({ name: "test/event/after" });

  return c.json({ message: "Done!" });
});
```

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [x] Added a [docs PR](https://github.com/inngest/website) that
references this PR
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

- Documented in inngest/website#873