If response signing fails when successfully validated req, throw (#737)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

If the request is signed, we expect that the response will also be
signed. This PR ensures that if signing the response fails, we throw a
`500 Internal Server Error` instead of silently failing.

This is to address issues where trust probes are performed but the
response is not signed.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Internal
- [x] Added unit/integration tests
- [x] Added changesets if applicable