Ensure support for `typescript@~5.7.0` (#760)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Ensures support for `typescript@~5.7.0` by adding a new PR check. See
[Announcing TypeScript 5.7 -
TypeScript](https://devblogs.microsoft.com/typescript/announcing-typescript-5-7/).

Ideally we would take advantage of the new
`rewriteRelativeImportExtensions` added in microsoft/TypeScript#59767
which would allow us to correctly import `.ts` extensions instead of
requiring `.js`, ready for better Deno/Bun support, as well as Node
support when running `.ts` files directly.

Unfortunately, `typescript<5.0.0` doesn't support the compiled output;
we'll have to wait for v4 to use that.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A KTLO
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

- https://devblogs.microsoft.com/typescript/announcing-typescript-5-7/
- microsoft/TypeSCript#59767