Revert "Try shifting bundler config - min 14 support"
This reverts commit 55eee183e5a22070a089f6b87ffabdb3a28285aa.

`@anthropic-ai/sdk` internally still requires `@types/node` too, a
dependency that we do not currently force on consumers. This
change would also requires that anyone using a `target` below
`"ES6"` in their `tsconfig.json` would need to upgade, which could
cause any number of additional build errors for them.

Changing our build outputs is a breaking change and we can't do it
here, so we have to do the more annoying task or manually copying
the types for now.