Fix `@inngest/test` automatic spying not accounting for `step.**` (#777)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

`@inngest/test` performs automatic mocking of the functions within the
`step` object. This change ensure we also recursively apply those mocks
such that `step.**()` is mocked, such as `step.ai.infer()`.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [ ] ~Added unit/integration tests~ N/A Using in another PR
- [x] Added changesets if applicable