Fix/named fn params (#798)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Broadens the types of `step.run()` to better recognize `Function` as
callable, therefore making the generic and allowing `Parameters<TFn>` to
extract correctly.

```ts
// This should now work
step.run("", function named() {});
```

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [x] Added unit/integration tests
- [x] Added changesets if applicable