Make adapter types an easily-accessible value (#871)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

We use an unexported `unique symbol` to store static types for AI
adapters. This works well as it can store typing without verbose
generics and hides the type when accessing the adapters e.g.
`adatper.*`.

However, a process that includes multiple versions of `@inngest/ai` can
then be problematic. This is often the case if `@inngest/ai` is a
sub-dependency of your project, there are multiple sub-dependencies on
`@inngest/ai`, or you have also installed your ownv version.

When this happens, the two `unique symbol`s are different and do not
match when comparing types.

Instead, we just let the type be visible, with a `~` prefix so it at
least goes to the bottom of autocomplete.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A KTLO
- [ ] ~Added unit/integration tests~ N/A KTLO
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Addressing build issues in inngest/agent-kit#87