Wait until messages have been flushed before setting `CLOSED` (#875)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Wait until messages have been flushed before setting `connection.state`
to `CLOSED`.

This means that the state is correctly `CLOSING` while performing a warm
shutdown.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A In progress cc @djfarrelly
- [ ] ~Added unit/integration tests~ N/A 
- [x] Added changesets if applicable